requirements.txt
pytest~=8.3.5
selene>=2.0.0rc9
allure-pytest>=2.13.3
python-dotenv>=1.0.0